LayoutTests/imported/w3c/web-platform-tests/digital-credentials/digital-credential-user-agent-allows-protocol.https.html

<!DOCTYPE html>
<title>Tests for DigitalCredential's userAgentAllowsProtocol() static method.</title>
<link rel="help" href="https://w3c-fedid.github.io/digital-credentials/#dom-digitalcredential-useragentallowsprotocol" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body></body>
<script type="module">

  const invalidIdentifiers = ["", " ", "a b", "a,b", "AB", "a.B", "a.b", " foo-bar", " foo-bar ", "foo-bar ", "fooBar"];

  promise_test(async (t) => {
    for (const invalidIdentifier of invalidIdentifiers) {
      const result = DigitalCredential.userAgentAllowsProtocol(invalidIdentifier);
      assert_equals(result, false, `Incorrect result for ${invalidIdentifier}.`);
    }
  }, "User agent does not allow invalid protocol identifiers.");

  const validPresentationProtocols = [
    "openid4vp-v1-unsigned",
    "openid4vp-v1-signed",
    "openid4vp-v1-multisigned",
    "org-iso-mdoc",
  ];

  for (const protocol of validPresentationProtocols) {
    test(() => {
      assert_true(
        DigitalCredential.userAgentAllowsProtocol(protocol),
        `Expected true for valid presentation protocol "${protocol}".`
      );
    }, `userAgentAllowsProtocol() returns true for "${protocol}".`);
  }

</script>
